rename: input외에 텍스트를 전부 포함하는 훅으로 이름 변경
