+++ src/hooks/useText.tsx
